Update qvm-ls invocation for R4
Unfortunately the new version is not backwards compatible with the old
syntax. I've put the old syntax first since I don't have a r3.2 machine
anymore to test if the old version is effectively forward-compatible in
a clean way, but this order should be safe.

The old way is preserved in branch `r3.2` regardless.